Rebased onto master, fixed errors
